etl_template/requirements.txt
flake8==4.0
psycopg2
pytest==6.2.5
python-dotenv==0.19.2
requests==2.27.1
elasticsearch==8.1.0
backoff